requirements.txt
joblib
scikit-learn
pymongo
python-dotenv
certifi
